requirements.txt
torch
torchvision
pandas
pillow
scikit-learn
tqdm